+++ index.js
@@ -46,6 +46,14 @@ module.exports = {
     // TODO: re-enable when an option for text children is available
     'react/jsx-one-expression-per-line': 'off',
 
+    // This is widely used in Backpack and internally and the inpact turning this on outweighs our perceived benefit of enforcing it
+    // https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/jsx-props-no-spreading.md
+    "react/jsx-props-no-spreading": 'off',
+
+    // This rule is pure about code style and the inpact turning this on outweighs our perceived benefit of enforcing it
+    // https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/static-property-placement.md
+    "react/static-property-placement": 'off',
+
     // Added 'to' as a specialLink property, which prevents react-router's
     // 'Link' component from triggering this rule.
     // See https://github.com/evcohen/eslint-plugin-jsx-a11y/issues/339